refactor: 코드리뷰 반영 Confetti 위치 변경
