+++ src/components/ReceiptResult/ReceiptResult.module.scss
@@ -53,7 +53,7 @@
 }
 
 .IconBtn {
-  width: 98.5px;
-  margin-top: 10px;
+  width: 6.15625rem;
+  margin-top: 0.625rem;
   margin-left: auto;
 }
